# Conflicts:
#	bishengir/lib/Dialect/HFusion/Transforms/DecomposeOpInterfaceImpl.cpp
#	bishengir/test/Conversion/HFusionToHIVM/hfusion-to-hivm.mlir

+++ bishengir/lib/Dialect/HIVM/Transforms/ConvertLayout/ShapeConversion/ConvertLayoutFromFractalsUtils.cpp
@@ -93,7 +93,7 @@ extractFractalToNDConversionParams(ArrayRef<OpFoldResult> currentShape,
 
   FractalToNDConversionParams params;
   params.fractalSize = *blockSizesResult;
-  params.batchIndexBias = currentShape.size() % 2;
+  params.batchIndexBias = static_cast<int>(currentShape.size()) % 2;
   LDBG("Batch index bias: " << params.batchIndexBias);
   if (params.batchIndexBias)
     params.batchDim = currentShape[0];
